# docs/source/reference/api-restapi/openapi.yml
openapi: 3.0.1
info:
  title: Radar REST API
  version: 0.0.0
servers:
- url: /
...
paths:
...
  /api/v1/entrypoints/{id}/plugins:
    get:
      tags:
      - Entrypoints
      summary: Retrieves the plugin snapshots for an entrypoint resource
      operationId: get_entrypoint_id_plugins_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the Entrypoint resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EntrypointPlugin'
    post:
      tags:
      - Entrypoints
      summary: Appends plugins to an entrypoint resource
      operationId: post_entrypoint_id_plugins_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the Entrypoint resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntrypointPluginMutableFields-load'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EntrypointPlugin'
      x-codegen-request-body-name: payload
  /api/v1/entrypoints/{id}/plugins/{pluginId}:
    get:
      tags:
      - Entrypoints
      summary: Retrieves a plugin snapshot for an entrypoint resource
      operationId: get_entrypoint_id_plugins_id_endpoint
      parameters:
      - name: pluginId
        in: path
        description: ID for the Plugin resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the Entrypoint resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntrypointPlugin'
    delete:
      tags:
      - Entrypoints
      summary: Removes a plugin from an entrypoint by ID
      operationId: delete_entrypoint_id_plugins_id_endpoint
      parameters:
      - name: pluginId
        in: path
        description: ID for the Plugin resource.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the Entrypoint resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/pluginParameterTypes/:
    get:
      tags:
      - PluginParameterTypes
      summary: Gets a list of all PluginParameterType resources
      operationId: get_plugin_parameter_type_endpoint
      parameters:
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: groupId
        in: query
        description: Filter results by the Group ID.
        schema:
          type: integer
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginParameterTypePage'
    post:
      tags:
      - PluginParameterTypes
      summary: Creates a PluginParameterType resource
      operationId: post_plugin_parameter_type_endpoint
      parameters:
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginParameterType-load'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginParameterType-dump'
      x-codegen-request-body-name: payload
...
  /api/v1/pluginParameterTypes/{id}:
    get:
      tags:
      - PluginParameterTypes
      summary: Gets a PluginParameterType resource
      operationId: get_plugin_parameter_type_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the PluginParameterType resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginParameterType-dump'
    put:
      tags:
      - PluginParameterTypes
      summary: Modifies a PluginParameterType resource
      operationId: put_plugin_parameter_type_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the PluginParameterType resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginParameterTypeMutableFields'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginParameterType-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - PluginParameterTypes
      summary: Deletes a PluginParameterType resource
      operationId: delete_plugin_parameter_type_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the PluginParameterType resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/plugins/:
    get:
      tags:
      - Plugins
      summary: Gets a list of all Plugin resources
      operationId: get_plugin_endpoint
      parameters:
      - name: search
        in: query
        description: Search terms for the query (* and ? wildcards).
        schema:
          type: string
      - name: groupId
        in: query
        description: Filter results by the Group ID.
        schema:
          type: integer
      - name: index
        in: query
        description: Index of the current page.
        schema:
          type: integer
      - name: pageLength
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginPage'
    post:
      tags:
      - Plugins
      summary: Creates a Plugin resource
      operationId: post_plugin_endpoint
      parameters:
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plugin-load'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plugin-dump'
      x-codegen-request-body-name: payload
...
  /api/v1/plugins/{id}:
    get:
      tags:
      - Plugins
      summary: Gets a Plugin resource
      operationId: get_plugin_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the Plugin resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plugin-dump'
    put:
      tags:
      - Plugins
      summary: Modifies a Plugin resource
      operationId: put_plugin_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the Plugin resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginMutableFields'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plugin-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - Plugins
      summary: Deletes a Plugin resource
      operationId: delete_plugin_id_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the Plugin resource.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
components:
  schemas:
...
    GroupRef:
      type: object
      properties:
        id:
          type: integer
          description: ID for the Group resource.
        name:
          type: string
          description: Name of the Group resource.
        url:
          type: string
          description: URL for accessing the full Group resource.
    UserRef-dump:
      type: object
      properties:
        id:
          type: integer
          description: ID for the User resource.
        username:
          type: string
          description: Username of the User resource.
        url:
          type: string
          description: URL for accessing the full User resource.
    TagRef-dump:
      type: object
      properties:
        id:
          type: integer
          description: ID for the Tag.
        name:
          type: string
          description: Name of the Tag.
        group:
          type: object
          description: Owner of the Tag.
          allOf:
          - $ref: '#/components/schemas/GroupRef'
        url:
          type: string
          description: URL for accessing the full Tag.
...
    EntrypointPlugin:
      type: object
      properties:
        id:
          type: integer
          description: ID for the Plugin resource.
        name:
          type: string
          description: Name of the Plugin resource.
        snapshotId:
          type: integer
          description: Snapshot ID for the Plugin resource.
        url:
          type: string
          description: URL for accessing the full Plugin snapshot.
        files:
          type: array
          items:
            type: object
            description: List of parameters for the entrypoint.
            allOf:
            - $ref: '#/components/schemas/EntrypointPluginFile'
    EntrypointPluginFile:
      type: object
      properties:
        id:
          type: integer
          description: ID for the PluginFile resource.
        filename:
          type: string
          description: Filename of the PluginFile resource.
        snapshotId:
          type: integer
          description: Snapshot ID for the PluginFile resource.
        url:
          type: string
          description: URL for accessing the full PluginFile snapshot.
...
    IdStatusResponse:
      type: object
      properties:
        id:
          type: array
          description: A list of integers identifying the affected resource(s).
          items:
            type: integer
        status:
          type: string
          description: The status of the request.
    EntrypointPluginMutableFields-load:
      type: object
      properties:
        plugins:
          type: array
          description: List of plugin files for the entrypoint.
          items:
            type: integer
...
    PluginParameterType-load:
      required:
      - group
      - name
      type: object
      properties:
        group:
          type: integer
          description: ID of the Group that will own the PluginParameterType resource.
        name:
          type: string
          description: Name of the PluginParameterType resource.
        description:
          type: string
          description: Description of the PluginParameterType resource.
        structure:
          type: object
          properties: {}
          description: Structure of the PluginParameterType resource.
    PluginParameterTypePage:
      type: object
      properties:
        index:
          type: integer
          description: Index of the current page.
        isComplete:
          type: boolean
          description: Boolean indicating if more data is available.
        totalNumResults:
          type: integer
          description: Total number of results.
        first:
          type: string
          description: URL to first page in results set.
        next:
          type: string
          description: URL to next page in results set.
        prev:
          type: string
          description: URL to last page in results set.
        data:
          type: array
          items:
            type: object
            description: List of PluginParameterType resources in the current page.
            allOf:
            - $ref: '#/components/schemas/PluginParameterType-dump'
    PluginParameterType-dump:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          description: ID for the PluginParameterType resource.
        snapshot:
          type: integer
          description: ID for the underlying PluginParameterType snapshot.
        group:
          type: object
          description: Group that owns the PluginParameterType resource.
          allOf:
          - $ref: '#/components/schemas/GroupRef'
        user:
          type: object
          description: User that created the PluginParameterType resource.
          allOf:
          - $ref: '#/components/schemas/UserRef-dump'
        createdOn:
          type: string
          description: Timestamp when the PluginParameterType resource was created.
          format: date-time
        lastModifiedOn:
          type: string
          description: Timestamp when the PluginParameterType resource was last modified.
          format: date-time
        latestSnapshot:
          type: boolean
          description: Whether or not the PluginParameterType resource is the latest
            version.
        hasDraft:
          type: boolean
          description: Whether a draft exists for the PluginParameterType resource.
        tags:
          type: array
          items:
            type: object
            description: Tags associated with the PluginParameterType resource.
            allOf:
            - $ref: '#/components/schemas/TagRef-dump'
        name:
          type: string
          description: Name of the PluginParameterType resource.
        description:
          type: string
          description: Description of the PluginParameterType resource.
        structure:
          type: object
          properties: {}
          description: Structure of the PluginParameterType resource.
    PluginParameterTypeMutableFields:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: Name of the PluginParameterType resource.
        description:
          type: string
          description: Description of the PluginParameterType resource.
        structure:
          type: object
          properties: {}
          description: Structure of the PluginParameterType resource.
...
    Plugin-load:
      required:
      - group
      - name
      type: object
      properties:
        group:
          type: integer
          description: ID of the Group that will own the Plugin resource.
        name:
          type: string
          description: Name of the Plugin resource.
        description:
          type: string
          description: Description of the Plugin resource.
    PluginPage:
      type: object
      properties:
        index:
          type: integer
          description: Index of the current page.
        isComplete:
          type: boolean
          description: Boolean indicating if more data is available.
        totalNumResults:
          type: integer
          description: Total number of results.
        first:
          type: string
          description: URL to first page in results set.
        next:
          type: string
          description: URL to next page in results set.
        prev:
          type: string
          description: URL to last page in results set.
        data:
          type: array
          items:
            type: object
            description: List of Plugin resources in the current page.
            allOf:
            - $ref: '#/components/schemas/Plugin-dump'
    Plugin-dump:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          description: ID for the Plugin resource.
        snapshot:
          type: integer
          description: ID for the underlying Plugin snapshot.
        group:
          type: object
          description: Group that owns the Plugin resource.
          allOf:
          - $ref: '#/components/schemas/GroupRef'
        user:
          type: object
          description: User that created the Plugin resource.
          allOf:
          - $ref: '#/components/schemas/UserRef-dump'
        createdOn:
          type: string
          description: Timestamp when the Plugin resource was created.
          format: date-time
        lastModifiedOn:
          type: string
          description: Timestamp when the Plugin resource was last modified.
          format: date-time
        latestSnapshot:
          type: boolean
          description: Whether or not the Plugin resource is the latest version.
        hasDraft:
          type: boolean
          description: Whether a draft exists for the Plugin resource.
        tags:
          type: array
          items:
            type: object
            description: Tags associated with the Plugin resource.
            allOf:
            - $ref: '#/components/schemas/TagRef-dump'
        name:
          type: string
          description: Name of the Plugin resource.
        description:
          type: string
          description: Description of the Plugin resource.
        files:
          type: array
          items:
            type: object
            description: Files associated with the Plugin resource.
            allOf:
            - $ref: '#/components/schemas/PluginFileRef'
    PluginFileRef:
      type: object
      properties:
        id:
          type: integer
          description: ID for the PluginFile resource.
        group:
          type: object
          description: Group that owns the PluginFile resource.
          allOf:
          - $ref: '#/components/schemas/GroupRef'
        url:
          type: string
          description: URL for accessing the full PluginFile resource.
        plugin:
          type: integer
          description: ID for the Plugin resource this file belongs to.
        filename:
          type: string
          description: Filename of the PluginFile resource.
    PluginMutableFields:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: Name of the Plugin resource.
        description:
          type: string
          description: Description of the Plugin resource.
...